Shattered GazePointer.h and some of GazePointer.cpp into individual class declarations.
