+++ Microsoft.Toolkit.Uwp.Input.Gaze/GazePointer.h
@@ -6,6 +6,9 @@
 
 #include "IGazeFilter.h"
 #include "GazeCursor.h"
+#include "GazeIsInvokableDelegate.h"
+#include "GazeInvokeTargetDelegate.h"
+#include "GazePointerState.h"
 
 using namespace Platform;
 using namespace Platform::Collections;
@@ -20,51 +23,6 @@ namespace Shapes = Windows::UI::Xaml::Shapes;
 
 BEGIN_NAMESPACE_GAZE_INPUT
 
-ref class GazeElement;
-ref class GazePointer;
-
-public ref class GazeApi sealed
-{
-public:
-    static property DependencyProperty^ IsGazeEnabledProperty { DependencyProperty^ get(); }
-    static property DependencyProperty^ IsGazeCursorVisibleProperty { DependencyProperty^ get(); }
-	static property DependencyProperty^ GazeCursorRadiusProperty { DependencyProperty^ get(); }
-
-    static property DependencyProperty^ GazeElementProperty { DependencyProperty^ get(); }
-
-    static property DependencyProperty^ FixationProperty { DependencyProperty^ get(); }
-    static property DependencyProperty^ DwellProperty { DependencyProperty^ get(); }
-    static property DependencyProperty^ DwellRepeatProperty { DependencyProperty^ get(); }
-    static property DependencyProperty^ EnterProperty { DependencyProperty^ get(); }
-    static property DependencyProperty^ ExitProperty { DependencyProperty^ get(); }
-
-	static property DependencyProperty^ MaxRepeatCountProperty { DependencyProperty^ get(); }
-
-    static bool GetIsGazeEnabled(Page^ page);
-    static bool GetIsGazeCursorVisible(Page^ page);
-	static int GetGazeCursorRadius(Page^ page);
-    static GazeElement^ GetGazeElement(UIElement^ element);
-    static TimeSpan GetFixation(UIElement^ element);
-    static TimeSpan GetDwell(UIElement^ element);
-    static TimeSpan GetDwellRepeat(UIElement^ element);
-    static TimeSpan GetEnter(UIElement^ element);
-    static TimeSpan GetExit(UIElement^ element);
-	static int GetMaxRepeatCount(UIElement^ element);
-
-    static void SetIsGazeEnabled(Page^ page, bool value);
-    static void SetIsGazeCursorVisible(Page^ page, bool value);
-	static void SetGazeCursorRadius(Page^ page, int value);
-    static void SetGazeElement(UIElement^ element, GazeElement^ value);
-    static void SetFixation(UIElement^ element, TimeSpan span);
-    static void SetDwell(UIElement^ element, TimeSpan span);
-    static void SetDwellRepeat(UIElement^ element, TimeSpan span);
-    static void SetEnter(UIElement^ element, TimeSpan span);
-    static void SetExit(UIElement^ element, TimeSpan span);
-	static void SetMaxRepeatCount(UIElement^ element, int value);
-
-	static GazePointer^ GetGazePointer(Page^ page);
-};
-
 // units in microseconds
 const int DEFAULT_FIXATION_DELAY = 400000;
 const int DEFAULT_DWELL_DELAY = 800000;
@@ -75,103 +33,8 @@ const int MAX_SINGLE_SAMPLE_DURATION = 100000;
 
 const int GAZE_IDLE_TIME = 2500000;
 
-public enum class GazePointerState
-{
-    Exit,
-
-    // The order of the following elements is important because
-    // they represent states that linearly transition to their
-    // immediate successors. 
-    PreEnter,
-    Enter,
-    Fixation,
-    Dwell,
-    //FixationRepeat,
-    DwellRepeat
-};
-
-ref struct GazeHistoryItem
-{
-    property UIElement^ HitTarget;
-    property int64 Timestamp;
-    property int Duration;
-};
-
-ref struct GazeTargetItem sealed
-{
-    property int ElapsedTime;
-    property int NextStateTime;
-    property int64 LastTimestamp;
-    property GazePointerState ElementState;
-    property UIElement^ TargetElement;
-	property int RepeatCount;
-	property int MaxRepeatCount;
-
-    GazeTargetItem(UIElement^ target)
-    {
-        TargetElement = target;
-    }
-
-    void Reset(int nextStateTime)
-    {
-        ElementState = GazePointerState::PreEnter;
-        ElapsedTime = 0;
-        NextStateTime = nextStateTime;
-		RepeatCount = 0;
-		MaxRepeatCount = GazeApi::GetMaxRepeatCount(TargetElement);
-    }
-};
-
-public ref struct GazePointerEventArgs sealed
-{
-    property UIElement^ HitTarget;
-    property GazePointerState PointerState;
-    property int ElapsedTime;
-    property TimeSpan ElapsedTimeSpan { TimeSpan get() { return *new TimeSpan{ 10 * ElapsedTime }; } }
-
-    GazePointerEventArgs(UIElement^ target, GazePointerState state, int elapsedTime)
-    {
-        HitTarget = target;
-        PointerState = state;
-        ElapsedTime = elapsedTime;
-    }
-};
-
-ref class GazePointer;
-public delegate void GazePointerEvent(GazePointer^ sender, GazePointerEventArgs^ ea);
-
-public delegate bool GazeIsInvokableDelegate(UIElement^ target);
-public delegate void GazeInvokeTargetDelegate(UIElement^ target);
-
-public ref class GazeInvokedRoutedEventArgs : public RoutedEventArgs
-{
-public:
-
-    property bool Handled;
-};
-
-public ref class GazeElement sealed : public DependencyObject
-{
-private:
-    static DependencyProperty^ const s_hasAttentionProperty;
-    static DependencyProperty^ const s_invokeProgressProperty;
-public:
-    static property DependencyProperty^ HasAttentionProperty { DependencyProperty^ get() { return s_hasAttentionProperty; } }
-    static property DependencyProperty^ InvokeProgressProperty { DependencyProperty^ get() { return s_invokeProgressProperty; } }
-
-    property bool HasAttention { bool get() { return safe_cast<bool>(GetValue(s_hasAttentionProperty)); } void set(bool value) { SetValue(s_hasAttentionProperty, value); } }
-    property double InvokeProgress { double get() { return safe_cast<double>(GetValue(s_invokeProgressProperty)); } void set(double value) { SetValue(s_invokeProgressProperty, value); } }
-
-    event GazePointerEvent^ GazePointerEvent;
-    event EventHandler<GazeInvokedRoutedEventArgs^>^ Invoked;
-
-    void RaiseGazePointerEvent(GazePointer^ sender, GazePointerEventArgs^ args) { GazePointerEvent(sender, args); }
-
-    void RaiseInvoked(Object^ sender, GazeInvokedRoutedEventArgs^ args)
-    {
-        Invoked(sender, args);
-    }
-};
+ref struct GazeTargetItem;
+ref struct GazeHistoryItem;
 
 public ref class GazePointer sealed
 {
