+++ Microsoft.Toolkit.Uwp.Input.Gaze/GazePointerState.cpp
@@ -1,0 +1,18 @@
+//Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license.
+//See LICENSE in the project root for license information.
+
+#include "pch.h"
+#include "GazePointerState.h"
+
+using namespace std;
+using namespace Platform;
+using namespace Windows::Foundation::Collections;
+using namespace Windows::Graphics::Display;
+using namespace Windows::UI;
+using namespace Windows::UI::ViewManagement;
+using namespace Windows::UI::Xaml::Automation::Peers;
+using namespace Windows::UI::Xaml::Hosting;
+
+BEGIN_NAMESPACE_GAZE_INPUT
+
+END_NAMESPACE_GAZE_INPUT
