+++ Microsoft.Toolkit.Uwp.Input.Gaze/Microsoft.Toolkit.UWP.Input.Gaze.vcxproj
@@ -209,19 +209,39 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="GazeApi.h" />
     <ClInclude Include="GazeCursor.h" />
+    <ClInclude Include="GazeElement.h" />
+    <ClInclude Include="GazeHistoryItem.h" />
+    <ClInclude Include="GazeInvokedRoutedEventArgs.h" />
+    <ClInclude Include="GazeInvokeTargetDelegate.h" />
+    <ClInclude Include="GazeIsInvokableDelegate.h" />
     <ClInclude Include="GazePointer.h" />
+    <ClInclude Include="GazePointerEvent.h" />
+    <ClInclude Include="GazePointerEventArgs.h" />
+    <ClInclude Include="GazePointerState.h" />
     <ClInclude Include="GazeSettingsHelper.h" />
     <ClInclude Include="GazeStats.h" />
+    <ClInclude Include="GazeTargetItem.h" />
     <ClInclude Include="IGazeFilter.h" />
     <ClInclude Include="OneEuroFilter.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="GazeApi.cpp" />
     <ClCompile Include="GazeCursor.cpp" />
+    <ClCompile Include="GazeElement.cpp" />
+    <ClCompile Include="GazeHistoryItem.cpp" />
+    <ClCompile Include="GazeInvokdeTargetDelegate.cpp" />
+    <ClCompile Include="GazeInvokedRoutedEventArgs.cpp" />
+    <ClCompile Include="GazeIsInvokableDelegate.cpp" />
     <ClCompile Include="GazePointer.cpp" />
+    <ClCompile Include="GazePointerEvent.cpp" />
+    <ClCompile Include="GazePointerEventArgs.cpp" />
+    <ClCompile Include="GazePointerState.cpp" />
     <ClCompile Include="GazeSettingsHelper.cpp" />
     <ClCompile Include="GazeStats.cpp" />
+    <ClCompile Include="GazeTargetItem.cpp" />
     <ClCompile Include="OneEuroFilter.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
