+++ Microsoft.Toolkit.Uwp.Input.Gaze/Microsoft.Toolkit.UWP.Input.Gaze.vcxproj.filters
@@ -31,6 +31,36 @@
     <ClCompile Include="GazeSettingsHelper.cpp">
       <Filter>Sources</Filter>
     </ClCompile>
+    <ClCompile Include="GazeApi.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazeElement.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazeHistoryItem.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazeInvokedRoutedEventArgs.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazeInvokdeTargetDelegate.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazeIsInvokableDelegate.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazePointerEventArgs.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazePointerState.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazeTargetItem.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="GazePointerEvent.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h">
@@ -52,6 +82,36 @@
       <Filter>Headers</Filter>
     </ClInclude>
     <ClInclude Include="GazeSettingsHelper.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeApi.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeElement.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeHistoryItem.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeInvokedRoutedEventArgs.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazePointerEventArgs.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazePointerState.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeTargetItem.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeInvokeTargetDelegate.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazeIsInvokableDelegate.h">
+      <Filter>Headers</Filter>
+    </ClInclude>
+    <ClInclude Include="GazePointerEvent.h">
       <Filter>Headers</Filter>
     </ClInclude>
   </ItemGroup>
